Move heredoc expansion from parse time to execute time
This is necessary to ensure variables have correct values in a list of
commands where earlier command modifies variables and later command
expands them in heredoc input.

Refactor heredoc parser by removing unnecessary local struct and
simplifying the line reading loop.
