+++ src/execute/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: amakinen <amakinen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/29 15:10:22 by josmanov          #+#    #+#             */
-/*   Updated: 2025/06/04 22:43:02 by amakinen         ###   ########.fr       */
+/*   Updated: 2025/06/04 22:57:06 by amakinen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,7 @@
 #include "libft.h"
 #include "status.h"
 #include "util.h"
+#include "word.h"
 
 #define TMP_FILE_PATH "/tmp"
 #define PROC_FD_PATH "/proc/self/fd/"
@@ -66,20 +67,29 @@ static t_status	execute_heredoc_create(int *writefd_out, int *readfd_out)
 }
 
 /*
-	Write heredoc lines to the temporary file.
+	Write heredoc lines to the temporary file. Expand the lines first if the
+	heredoc was not quoted.
 
 	Write doesn't need a null terminator, so we can insert the newline in its
 	place and not need a separate write call. Put the terminator back afterwards
 	in case some other code tries to process the string after this function.
 */
-static t_status	execute_heredoc_write(int fd, struct s_ast_command_word *lines)
+static t_status	execute_heredoc_write(int fd, bool quoted,
+	struct s_ast_command_word *lines)
 {
+	t_status					status;
 	struct s_ast_command_word	*current;
 	size_t						len;
 
 	current = lines;
 	while (current)
 	{
+		if (!quoted)
+		{
+			status = word_heredoc_line(&current->word);
+			if (status != S_OK)
+				return (status);
+		}
 		len = ft_strlen(current->word);
 		current->word[len] = '\n';
 		if (!util_write_all(fd, current->word, len + 1))
@@ -112,7 +122,8 @@ t_status	execute_redirect_heredoc(struct s_ast_redirect *redirect,
 	status = execute_heredoc_create(&writefd, &readfd);
 	if (status != S_OK)
 		return (status);
-	status = execute_heredoc_write(writefd, redirect->heredoc_lines);
+	status = execute_heredoc_write(writefd, redirect->heredoc_quoted,
+			redirect->heredoc_lines);
 	if (close(writefd) < 0)
 		if (status == S_OK)
 			status = status_err(S_COMM_ERR, "execute_heredoc",
